+++ _test_go_script.py
@@ -9,6 +9,7 @@ load_dotenv()
 
 OCTO_ACC_NO = os.getenv('OCTO_ACC_NO')
 OCTO_API_KEY = os.getenv('OCTO_API_KEY')
+SMART_PLUG_IP = os.getenv('SMART_PLUG_IP')
 
 logging.getLogger().setLevel(logging.DEBUG)
 
@@ -16,16 +17,16 @@ logging.debug("starting")
 
 def switch_shelly_on_and_alert(price: Price):
     logging.info(f"Price is cheap: {price}")
-    requests.get("http://192.168.1.235/relay/0?turn=on")
+    requests.get(f"http://{SMART_PLUG_IP}/relay/0?turn=on")
 
 def switch_shelly_off_and_alert(price: Price):
     logging.info(f"Price is expensive: {price}")
-    requests.get("http://192.168.1.235/relay/0?turn=off")
+    requests.get(f"http://{SMART_PLUG_IP}/relay/0?turn=off")
 
-runner.run_octopus_agile_tariff_schedule(
-    prices_to_include=10,
+runner.run_octopus_go_tariff_schedule(
     action_when_cheap=switch_shelly_on_and_alert,
     action_when_expensive=switch_shelly_off_and_alert,
     api_key=OCTO_API_KEY,
-    account_number=OCTO_ACC_NO
+    account_number=OCTO_ACC_NO,
+    is_intelligent=True
 )