+++ CHANGELOG.md
@@ -19,3 +19,23 @@ Format: [https://gist.github.com/juampynr/4c18214a8eb554084e21d6e288a18a2c](http
 
 - Made the initial schedule job run if the current time is within 15 minutes of the job execution time
 - Users now have to provide an API key for Octopus runners, this is to get the accurate tariff, previously assumed a single tariff that soon became unavailable.
+
+## [0.0.6] - 2025-01-03
+
+### Added
+
+- Intelligent Go and Regular Go tariff support
+
+### Changed
+
+### Fixed
+
+## [0.0.7] - 2025-01-03
+
+### Added
+
+### Changed
+
+- Removed interpolations, jobs for Go and Intelligent Go now only trigger during price boundaries
+
+### Fixed